requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.0.1
bleach==6.2.0
bx_django_utils==84
bx_py_utils==108
certifi==2025.1.31
charset-normalizer==3.4.1
colorlog==6.9.0
dependency-injector==4.46.0
Django==4.2.9
django-access-tokens==0.9.2
django-dotenv==1.4.2
django-environ==0.12.0
django-filter==23.5
django-redis==5.3.0
django-tools==0.56.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
ecdsa==0.19.1
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
frozenlist==1.5.0
gunicorn==23.0.0
icdiff==2.0.7
idna==3.10
inflection==0.5.1
kavenegar==1.1.1
multidict==6.3.2
packaging==24.2
persian-tools==0.0.11
persiantools==5.2.0
phonenumbers==8.13.22
pprintpp==0.4.0
propcache==0.3.1
pyasn1==0.6.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-slugify==8.0.1
python-stdnum==1.20
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
sendgrid==6.11.0
sentry-sdk==1.40.6
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
text-unidecode==1.3
tokenizer==3.4.5
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
wheel==0.45.1
yarl==1.19.0
